[CONF] Add popcount instruction
Although Penryn processor did not have popcount instruction, adding it here works.

This is e. g. required for OpenCV, since even when compiling from source, the detection seems to fail (in OpenCV 4.5.2_4 built with homebrew) and popcount is always recognized also in case it is not added in the CPU configuration. This commit fixes this issue.